카드 구조체를 Card.swift 파일로 분리
