+++ CardGameApp/CardGameApp/ViewController.swift
@@ -17,42 +17,6 @@ class ViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         setView()
-        let blackjackCard = BlackjackCard(suit: BlackjackCard.Suit(rawValue: "♥️"), rank: BlackjackCard.Rank(rawValue: 12))
-        blackjackCard.printCard()
-    }
-    
-    /*
-     - 카드 특성상 모양과 숫자의 모든 case를 알고 있고, 추가, 변경, 삭제되지 않으므로 열거형으로 정의
-     - 숫자의 경우 첫 번째 case에 Int 타입의 rawValue를 설정하여 1씩 증가시키기 편리함
-     - 정의한 타입이 카드에 내포되는 것을 파악하기 쉽도록 중첩하여 구현
-     */
-    private struct BlackjackCard {
-        enum Suit: Character {
-            case spades = "♠️", hearts = "♥️", diamonds = "♦️", clubs = "♣️"
-        }
-
-        enum Rank: Int {
-            case one = 1, two, three, four, five, six, seven, eight, nine, ten, eleven, twelve, thirteen
-        }
-
-        let suit: Suit?, rank: Rank?
-
-        func printCard() {
-            var output = ""
-            if let suit = suit {
-                output += String(suit.rawValue)
-            }
-            if let rank = rank {
-                let rankValue = rank.rawValue
-                if rankValue == 1 || rankValue > 10 {
-                    let ranks = [1: "A", 11: "J", 12: "Q", 13: "K"]
-                    output += ranks[rankValue]!
-                } else {
-                    output += String(rankValue)
-                }
-            }
-            print(output)
-        }
     }
     
     private func specifiedStackView() -> UIStackView {
